Root to leaf path sum
Given a Binary Tree and a sum s, your task is to check whether there is a root to leaf path in that tree with the following sum. You are required to complete the function hasPathSum. You should not read any input from stdin/console. There are multiple test cases. For each test case, this method will be called individually.