Binary Swap
You are given two binary strings A and B of equal length. You have a type of operation in which you can swap any two elements of string B. Your task is to find the minimum number of operations required to convert string B into string A. If it is not possible, print -1.

Input Format:

The first line contains string A.
The second line contains string B.

Output Format:
Print the minimum number of operations required to convert string B into string A. If it is not possible, print -1.